requirements-bot.txt
python-telegram-bot==20.8
PyYAML>=6.0
